log extension calls
Put timestamps in a file in a shared directory when the extension gets
called. I'll need to add a way of reading/exporting this file in the
main app later. Doing this so I can find out if Safari will
automatically reload rules when they change at all, or if it's always a
manual process. Since, you know, the docs don't mention this at all &
there's no API for notifying Safari rules have changed.