isolate tests better
this is dirty, but I don't know of a cleaner way to do it for Cocoa:
since tests run "in" the app, any behavior triggered by app starting
runs in the tests, which can be irritating.

Current solution: pass an ENV var when running tests, and look for it
before running some code.