try to address core data concurrency better
so it seems like maybe it's best practice to pretty much always use a
"child" context. and when you save a child context, you also need to
save the master, and you need that stuff to be protected in blocks. (do
those blocks implicitly lock the contexts in question? i hope so.)